refactor(playground): align with react-effects-policy
`docs/internal/react-effects-policy.md` says component code should
avoid raw \`useEffect\` and reach for purpose-named hooks from
\`app/hooks/effects.ts\` instead.

- canvas.tsx: extract the \`@viz-js/viz\` SVG render + selection
  highlight effect into a new \`useCanvasRender\` hook. As a side
  benefit the highlight no longer pays the layout cost when only
  the selection changes — two effects instead of one shared body.
- run-for-real-modal.tsx: swap the inline Escape-key effect for
  \`useDocumentEvent("keydown", …)\` from the existing primitives.
- persist.ts: drop the \`isFirstRender\` guard. It saved a single
  no-op localStorage write on mount, and the guard explanation was
  longer than the code it guarded.
